refactor(module): Inline DSL, remove Binder & BindingChain, make Node Bindable
