+++ app/src/androidTest/java/com/harrytmthy/stitch/StitchVsKoinBenchmark.kt
@@ -23,6 +23,7 @@ import androidx.benchmark.junit4.measureRepeated
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.harrytmthy.stitch.api.ScopeRef
 import com.harrytmthy.stitch.api.Stitch
+import com.harrytmthy.stitch.api.bind
 import com.harrytmthy.stitch.api.module
 import com.harrytmthy.stitch.api.named
 import com.harrytmthy.stitch.api.scope
