+++ stitch/api/stitch.api
@@ -1,15 +1,5 @@
-public final class com/harrytmthy/stitch/api/Binder {
-	public fun <init> (Z)V
-	public final fun factory (Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
-	public static synthetic fun factory$default (Lcom/harrytmthy/stitch/api/Binder;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
-	public final fun scoped-OHUMBuI (Ljava/lang/String;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
-	public static synthetic fun scoped-OHUMBuI$default (Lcom/harrytmthy/stitch/api/Binder;Ljava/lang/String;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
-	public final fun singleton (Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;ZLkotlin/jvm/functions/Function1;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
-	public static synthetic fun singleton$default (Lcom/harrytmthy/stitch/api/Binder;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;ZLkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
-}
-
-public final class com/harrytmthy/stitch/api/Binder$BindingChain {
-	public final fun bind (Ljava/lang/Class;)Lcom/harrytmthy/stitch/api/Binder$BindingChain;
+public abstract interface class com/harrytmthy/stitch/api/Bindable {
+	public abstract fun bind (Ljava/lang/Class;)Lcom/harrytmthy/stitch/api/Bindable;
 }
 
 public final class com/harrytmthy/stitch/api/Component {
@@ -19,9 +9,14 @@ public final class com/harrytmthy/stitch/api/Component {
 	public static synthetic fun lazyOf$default (Lcom/harrytmthy/stitch/api/Component;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lcom/harrytmthy/stitch/api/Scope;ILjava/lang/Object;)Lkotlin/Lazy;
 }
 
-public abstract class com/harrytmthy/stitch/api/Module {
-	public fun <init> (Z)V
-	public final fun getBinder ()Lcom/harrytmthy/stitch/api/Binder;
+public final class com/harrytmthy/stitch/api/Module {
+	public fun <init> (ZLkotlin/jvm/functions/Function1;)V
+	public final fun factory (Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;)Lcom/harrytmthy/stitch/api/Bindable;
+	public static synthetic fun factory$default (Lcom/harrytmthy/stitch/api/Module;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcom/harrytmthy/stitch/api/Bindable;
+	public final fun scoped-OHUMBuI (Ljava/lang/String;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;)Lcom/harrytmthy/stitch/api/Bindable;
+	public static synthetic fun scoped-OHUMBuI$default (Lcom/harrytmthy/stitch/api/Module;Ljava/lang/String;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcom/harrytmthy/stitch/api/Bindable;
+	public final fun singleton (Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;ZLkotlin/jvm/functions/Function1;)Lcom/harrytmthy/stitch/api/Bindable;
+	public static synthetic fun singleton$default (Lcom/harrytmthy/stitch/api/Module;Ljava/lang/Class;Lcom/harrytmthy/stitch/api/Qualifier;ZLkotlin/jvm/functions/Function1;ILjava/lang/Object;)Lcom/harrytmthy/stitch/api/Bindable;
 }
 
 public final class com/harrytmthy/stitch/api/ModuleKt {
